feat: add timeout promise to prevent hanging actions
