+++ libs/web3-api-provider/src/index.ts
@@ -1,10 +1,11 @@
 // Copyright 2022 @webb-tools/
 // SPDX-License-Identifier: Apache-2.0
 
+export * from './ext-provider';
+export * from './utils';
+export * from './webb-provider';
 export * from './webb-provider/bridge-api';
 export * from './webb-provider/chain-query';
 export * from './webb-provider/relayer-manager';
 export * from './webb-provider/vanchor-actions';
-export * from './webb-provider';
 export * from './webb-provider/wrap-unwrap';
-export * from './ext-provider';
